requirements.txt
# Web framework and server
fastapi==0.115.6
uvicorn[standard]==0.34.0

# File upload support — required by FastAPI for multipart/form-data parsing
python-multipart==0.0.20

# OCR engine — uses PaddleOCR's neural network models via ONNX Runtime.
# Much more accurate than Tesseract on stylized labels, and lighter than full PaddlePaddle.
rapidocr-onnxruntime

# Image loading and preprocessing before OCR
Pillow==11.1.0

# Server-side HTML templating for the form/results UI
Jinja2==3.1.5

# Testing
pytest==8.3.4
httpx==0.28.1
